requirements.txt
PyYAML
mpi4py
numpy
# tensorflow==1.10.1
pillow
gym
stable-baselines>=2.4.0
pytablewriter
# -e git+git@github.com:tawnkramer/donkey.git@faa4a32b2aff452c6b46474029ea9e4b3168843d#egg=donkeycar
opencv-python
cloudpickle
pygame
joblib
tqdm
